requirements.txt
streamlit
faiss-cpu
speechrecognition
pydub
transformers
torch
